requirements.txt
pandas
numpy
scikit-learn
streamlit
plotly
joblib
google-genai
python-dotenv